DS-2970 Add support for older xgboost outputs
Add try catch logic into the predict.GradientBoost method to attempt to
salvage an older computed xgboost output to be compatible with the newer
predict method in xgboost. Added an rds file which contains an older
such output in the test directory for use in unit tests to check that
the output can be successfully imported.